requirements.txt
celery==5.0.5
channels==3.0.4
cloudinary==1.26.0
crypto==1.4.1
django-ckeditor==6.1.0
django-cors-headers==3.2.1
django-filter==2.4.0
django-location-field==2.1.0
django-sslserver==0.22
djangorestframework>=3.13.1,!=3.14.0
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
dnspython==1.16.0
drf-yasg==1.20.0
gunicorn==20.1.0
haversine==2.5.1
mongoengine==0.23.1
pillow==8.1.0
pip-chill==1.0.1
psycopg2-binary==2.8.6
pyopenssl==22.0.0
pyrebase4==4.5.0
qrcode==7.3.1
redis==3.5.3
service-identity==21.1.0